backend/requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
openai==1.3.0
google-cloud-speech==2.21.0
pydub==0.25.1
python-dotenv==1.0.0
httpx==0.24.1
requests==2.31.0
elevenlabs==0.2.26
